+++ src/Controller/DefaultController.php
@@ -47,7 +47,7 @@ final class DefaultController extends BaseController
 
     private function getDbStats(): array
     {
-        $userService = $this->container->get('user_service');
+        $userService = $this->container->get('find_user_service');
         $taskService = $this->container->get('task_service');
         $noteService = $this->container->get('find_note_service');
