+++ src/Controller/User/Base.php
@@ -7,12 +7,18 @@ namespace App\Controller\User;
 use App\Controller\BaseController;
 use App\Exception\User;
 use App\Service\User\UserService;
+use App\Service\User\Find;
 
 abstract class Base extends BaseController
 {
     protected function getUserService(): UserService
     {
         return $this->container->get('user_service');
+    }
+
+    protected function getFindUserService(): Find
+    {
+        return $this->container->get('find_user_service');
     }
 
     protected function checkUserPermissions(int $userId, int $userIdLogged): void
