+++ src/Service/User/UserService.php
@@ -9,32 +9,6 @@ use Firebase\JWT\JWT;
 
 final class UserService extends Base
 {
-    public function getUsersByPage(
-        int $page,
-        int $perPage,
-        ?string $name,
-        ?string $email
-    ): array {
-        if ($page < 1) {
-            $page = 1;
-        }
-        if ($perPage < 1) {
-            $perPage = self::DEFAULT_PER_PAGE_PAGINATION;
-        }
-
-        return $this->userRepository->getUsersByPage(
-            $page,
-            $perPage,
-            $name,
-            $email
-        );
-    }
-
-    public function getAll(): array
-    {
-        return $this->userRepository->getAll();
-    }
-
     public function getOne(int $userId): object
     {
         if (self::isRedisEnabled() === true) {
